Add CORS Headers
Adds headers to allow cross origin usage of the mjpeg stream. These headers are required when interacting with the mjpeg image using canvas or webgl. You can add the headers in express, but this seems like a common enough use case that the proxy server itself should do this automatically.